LayoutTests/fast/loader/cancel-load-during-port-block-timer-expected.txt

CONSOLE MESSAGE: Not allowed to use restricted network port 22: http://127.0.0.1:22/
If this does crash, the test has passed.
